pyproject.toml
[project]
name = "fastapi-p31"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.30.0",
    "fastapi>=0.119.0",
    "flake8>=7.3.0",
    "greenlet>=3.2.4",
    "isort>=7.0.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.3",
    "python-jose>=3.5.0",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]

[tool.alembic]
script_location = "%(here)s/migrations"
version_locations = ["migrations/versions"]
prepend_sys_path = "."

[tool.isort]
profile = "hug"
skip = ["build", ".venv", "migrations"]
src_paths = ["isort", "test"]